Updated build platforms.
Wow, blast from the past. Back in the day, I developed BBug on a
PowerPC machine (an iBook), so I built the Debug version for PPC,
and made a PPC/i386 Release version. I've updated the build so that
the Debug version builds for the native machine architecture, and
the Release version is a 32-bit/64-bit universal binary.