Hide spinner when deserializing from nib.
For some reason, [self setHidden:YES] does not work in the spinner's
init method, possibly because it's being loaded from a nib file.

Although I thought the spinner wasn't working, it actually is. I think
maybe the data file is typically parsed and loaded quickly enough that
the spinner cannot be seen.