Moved all weather information into the status menu.
It occurred to me tonight that the "modern" way of handling this
would be to write a Dashboard widget -- because honestly, who keeps
a big widget window on their desktops anymore? But updating the
app to be a Dashboard widget would take a lot of work (and probably
require a rewrite in JavaScript), which I'm _not_ going to do right
now. So I decided to take the route of Apple's status menu items
(e.g., the battery indicator), and put the data _into_ the menu,
rather than having a separate window.

So it occurs to me that I did all that GUI work for naught, because
I basically just moved the old Dock menu into the status menu. _But_
I did write the code for the new menu a _lot_ better than the old
menu code was written, so that's definitely a big improvement.

This commit also removes the GUI library and code I was using from
the Xcode project...but the GUI framework itself was pretty cool,
so I'm leaving it in the repo in case I want to use it again.