Added a Next Update: item to the menu.
This was, naturally, included in the Dock menu version of this app.

I also added a category for NSTimer. -[NSTime fireDate] returns the
current time _if_ the timer just fired, but we always want the
_real_ next time, so I added a method to get that date. This was
present in the old version of the source code, from which I shamelessly
stole the implementation.