Added some tracking for a crash.
When the feed updates around midnight (usually around 11:45 PM), Bbug
has been crashing with a segfault. I have no idea what the cause is,
so I'm adding some code to track it down, as well as a copy of a data
file on which the app crashed.

I also added a new build configuration that is set to use the local
copy of the data file, rather than fetching from the Internet.