Correct weather feed timestamp in UI
The reasoning is explained in a comment in the code, but I'll
reiterate here for my notes. The problem is that the timestamp in
the weather feed is either (a) wrong by an hour, or (b) expressed in
an unexpected time zone. For example, say it's 8:02 PM and the feed
was just updated; it'll have the timestamp "2100", rather than the
expected "2000". Maybe this is semantically correct for some reason,
but it's definitely confusing for the user.

I'm going to contact Bucknell's weather guy to find out what the
timestamp really means, but until then I'll leave this chunk of code
in place.