Bump version to 2.0b2
Bucknell ISR (excuse me, _LIT_) is still testing this app, although I
have yet to receive any bug reports or feedback. It might be uncouth
to push an update in the middle of testing, but I've made some changes
since 2.0b1 that I'd really like to see in the final 2.0.