Fix timestamp glitch
_Sometimes_ the timestamp in the feed _is_ correct, and we shouldn't
try to correct it. There's not a lot of rhyme or reason as to when,
as far as I can tell, so if the raw timestamp is from within the
last hour, don't try to correct it.

I'd theorized that the timestamp was somehow correcting for DST
using the _old_ DST rules, but I don't think that's the case, since
under the _old_ rules, DST wouldn't've ended yet anyway.

This will probably break at some point, in which case I'm probably
just going to throw my hands in the air and say fuck it.

Closes #10.