Don't correct data file timestamp by default
This date correction mumbo-jumbo for the data file timestamp is
cumbersome, annoying, and error-prone, so by default, I'm just not
going to bother doing it. I left the correction code in, and it can
be enabled with a compile-time #define, but by default I'm going to
compile the application without the correction code.