Fix a bug in relative date additions
When checking to see if a date was on or before yesterday, or on or
after tomorrow, I only used the day component. This would make a
date fail if it had the exact same day (say, a month in the future).
This change takes the year and month into account.