Correct timestamp in application controller
The timestamp is sometimes "ahead" by an hour, meaning that the "Last
Updated" time is actually in the future. I'm not sure why; my hypothesis
is that the feed doesn't compensate for daylight saving time, but I
haven't found a satisfactory way to adjust for this. It makes sense to
do it in the application controller, anyway, so the adjustment is
pretty simple: if the data feed timestamp is in the future, display
it as being "back in time" by an hour instead.