Upgrade project to Xcode 4
This brings some pretty big changes to the app. I now use LLVM to compile,
and I target 10.6 and above; I'd like to maintain support for 10.4+, but
doing so is kind of a pain in the ass, and I don't have a machine to test
such a build on anyway.

I removed some old targets and build scripts that I no longer used; I'm
not sure how to fit them into Xcode 4, and they're not really needed
anyway.