Update build number
Aha! Thanks to [1] and [2], I finally have it figured out: a build
number may be up to three period-separated integers, but that does
_not_ make the number a float; rather, each integer is treated
separately. Thus, 10.1 is not the same as 10.10 (the former is
"less than" the latter), so I can basically encode the version
number in the first integer, and the build number in the second.
Simple!

  [1] http://bit.ly/19rfrbj
  [2] http://bit.ly/1htNApW