Commit v1.1 to repo
v1.1 was the only released version of BucknellBug. v2.0 (and ultimately
v2.1) was started to fix significant problems with the app, including
the UI. Unfortunately, v2.0, while completed, was never released. In the
meantime, Bucknell shut down its weather station, so this version and
subsequent versions don't even work.

This commit is the source code as released by Bucknell, preserved for
posterity. BucknellBug for Mac was originally developed in 2006 and
released in early 2007, so this code likely doesn't even compile on
modern Macs anymore.