Update weather menu using success and failure handlers
Due to the way BucknellBug was previously architected, there is now some
wonkiness in how the last updated entry on the menu is calculated; the
app reports failure, but the menu is still updated. I think this is
actually correct behavior and should work fine with the Dark Sky API
service, but it may be a bit weird, so I should keep an eye out for
that.