Only bump build number when archiving the project
Unfortunately, it seems that the plist value is not updated until after
the application is built and archived. This probably isn't a huge
problem, as it simply means the _next_ build will pick up the build
number, but I'd like it to happen for the version being built; I'll have
to work on solving that problem.

According to everything I've read on the Internet, this should _not_ be
happening, so I'm a bit perplexed as to the cause. However, [1] suggests
that perhaps it may be happening because the archive pre-actions are run
_after_ building, so this really needs to happen as a build pre-action,
but then it will occur on _every_ build, which I don't want... So I'm
not sure how best to deal with that. ([1] suggests creating a new
target, but I'd really like to avoid going that route.)

  [1] http://stackoverflow.com/a/5834892/28804