Grab rainfall amounts from daily data
For now, we can grab the rainfall amounts from the daily data. This
isn't really 0.0 inches in the past hour, but it is, at least, better
than showing 0.0 all the time. However, the daily "precipitation"
amount, as pulled from Dark Sky, does not necessarily reflect _rainfall_
amounts -- it may be snow, sleet, or hail, too. I could get fancy and
handle this, but for now, I'll treat it all as rain.

Refers to #13.