Increment build number only when product is archived
It was kind of annoying having the build number incremented every time I
built the product (even just for a quick test).

The weird thing is that the archive copy has the _previous_ build
number, _not_ what it was incremented to...but that should be okay. I
think. Whatever, works.